[#1910] Keep the MaxSpendMode FFI mapping in the Rust bridge layer
